chore(core): break circular dependency from getAuthTokenFromUserId
